refactor: . pull addPostponeButton() up
