+++ src/Renderer/QueryRenderer.ts
@@ -14,11 +14,10 @@ import type { TasksEvents } from '../Obsidian/TasksEvents';
 import type { GroupDisplayHeading } from '../Query/Group/GroupDisplayHeading';
 import type { TaskGroups } from '../Query/Group/TaskGroups';
 import type { QueryResult } from '../Query/QueryResult';
-import { postponeButtonTitle, shouldShowPostponeButton } from '../Scripting/Postponer';
+import { shouldShowPostponeButton } from '../Scripting/Postponer';
 import { TasksFile } from '../Scripting/TasksFile';
 import { DateFallback } from '../Task/DateFallback';
 import type { Task } from '../Task/Task';
-import { PostponeMenu } from '../ui/Menus/PostponeMenu';
 import { QueryResultsRenderer } from './QueryResultsRenderer';
 import { TaskLineRenderer, createAndAppendElement } from './TaskLineRenderer';
 
@@ -402,33 +401,5 @@ class QueryRenderChild extends QueryResultsRenderer {
         if (!shortMode) {
             backLink.append(')');
         }
-    }
-
-    private addPostponeButton(listItem: HTMLElement, task: Task, shortMode: boolean) {
-        const amount = 1;
-        const timeUnit = 'day';
-        const buttonTooltipText = postponeButtonTitle(task, amount, timeUnit);
-
-        const button = createAndAppendElement('a', listItem);
-        button.addClass('tasks-postpone');
-        if (shortMode) {
-            button.addClass('tasks-postpone-short-mode');
-        }
-        button.title = buttonTooltipText;
-
-        button.addEventListener('click', (ev: MouseEvent) => {
-            ev.preventDefault(); // suppress the default click behavior
-            ev.stopPropagation(); // suppress further event propagation
-            PostponeMenu.postponeOnClickCallback(button, task, amount, timeUnit);
-        });
-
-        /** Open a context menu on right-click.
-         */
-        button.addEventListener('contextmenu', async (ev: MouseEvent) => {
-            ev.preventDefault(); // suppress the default context menu
-            ev.stopPropagation(); // suppress further event propagation
-            const menu = new PostponeMenu(button, task);
-            menu.showAtPosition({ x: ev.clientX, y: ev.clientY });
-        });
     }
 }
